requirements.txt
aiogram==3.18.0
loguru==0.7.3
python-dotenv==1.0.1
requests==2.28.2
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
aiohttp
googletrans
langdetect